Реструктуризация проекта
