+++ Dockerfile
@@ -26,7 +26,7 @@ RUN pip install --upgrade pip \
 # 4. Запускаем UV, чтобы он поставил всё из pyproject.toml/uv.lock
 RUN uv sync
 
-# 5. Теперь копируем весь код бота (main.py, settings.py и т.д.)
+# 5. Теперь копируем весь код бота (app/ и другие файлы)
 COPY . /app
 
 # 8. Копируем скрипт миграции и делаем его исполняемым
@@ -36,5 +36,5 @@ RUN chmod +x /app/entrypoint.sh
 # 9. Скрипт-мигратор запускается первым, затем передаёт управление UV
 ENTRYPOINT ["/app/entrypoint.sh"]
 
-# 6. Команда по умолчанию: запускаем через UV (он найдёт ваши зависимости и сам выполнит main.py)
-CMD ["uv", "run", "main.py"]
+# 6. Команда по умолчанию: запускаем через UV (он найдёт ваши зависимости и сам выполнит app/main.py)
+CMD ["uv", "run", "app/main.py"]
