+++ app/db/models.py
@@ -2,14 +2,12 @@ from datetime import datetime, date
 from typing import Optional, List
 
 from sqlalchemy import (
-    BigInteger, Boolean, Column, Date, DateTime, ForeignKey, Integer, String, text
+    BigInteger, Boolean, Date, DateTime, String, text
 )
-from sqlalchemy.ext.declarative import declarative_base
 
 from sqlalchemy.orm import Mapped, mapped_column
 
-# SQLAlchemy 2.0 Declarative Base
-Base = declarative_base()
+from app.core.database import Base
 
 class User(Base):
     __tablename__ = 'users'
